Add bandwidth tests for xcore.ai and regression tests
Also, removed duplication in test scripts
