+++ tests/test_connection.py
@@ -8,40 +8,26 @@ import time
 import pathlib
 import subprocess
 from hardware_test_tools import XcoreApp
+from CollectFailures import CollectFailures
 
 
 @pytest.mark.parametrize('protocol', ['UDP', 'TCP'])
 @pytest.mark.parametrize('library', ['LWIP'])
 @pytest.mark.parametrize('processes', [1, 2])
-def test_connection(request, protocol, library, processes):
+@pytest.mark.parametrize('target', ['XMS0020'])
+def test_connection(request, protocol, library, processes, target):
     dut_ip = '192.168.200.198'
     dut_ports_per_proc = processes  # Number of Ports and processes parameterised the same for simplicity
 
     tester = RunXtcp(processes, dut_ports_per_proc, protocol)
     tester.setup(request)
 
-    tester.run_test(0.002, 'EXPLORER', dut_ip, 'ETH', library)
+    tester.run_test(0.002, target, dut_ip, 'ETH', library)
 
     tester.check_xrun()
     tester.check_python()
 
     assert len(tester.failures()) == 0
-
-
-class CollectFailures():
-    def __init__(self):
-        self._failures = []
-
-    def record_failure(self, failure_reason):
-        # Append a newline if there isn't one already
-        if not failure_reason.endswith('\n'):
-            failure_reason += '\n'
-        self._failures.append(failure_reason)
-        print("Failure reason: {}".format(failure_reason))
-        self.result = False
-
-    def failures(self):
-        return self._failures
 
 
 class RunXtcp(CollectFailures):
@@ -72,8 +58,8 @@ class RunXtcp(CollectFailures):
         else:  # weekend
             self.packets = 100
 
-    def run_test(self, delay, device, ip, interface, library):
-        setup = f'{self.processes}_{self.ports}_{self.protocol}_{device}_{interface}'
+    def run_test(self, delay, target, ip, interface, library):
+        setup = f'{self.processes}_{self.ports}_{self.protocol}_{target}_{interface}'
         binary = pathlib.Path(f'xtcp_bombard_{library.lower()}/bin/{setup}/xtcp_bombard_{library.lower()}_{setup}.xe')
 
         assert binary.exists(), 'Found test binary'
