+++ tests/xtcp_bombard_lwip/CMakeLists.txt
@@ -43,24 +43,24 @@ set(COMPILER_FLAGS_COMMON   -O3
                             -DDEBUG_PRINT_ENABLE=1
                             -DDEBUG_PRINT_ENABLE_LIB_XTCP=1
                             -DXASSERT_ENABLE_DEBUG=1
-                            -DTEST_BOARD_SUPPORT_BOARD=XK_ETH_XU316_DUAL_100M
+                            -DTEST_BOARD_SUPPORT_BOARD=XK_ETH_XU316_DUAL_100M # XMS0020 + XMS0019 + XMS0018?
                             ${PHY_FLAGS})
 
 # 1 process, 1 port
-set(APP_COMPILER_FLAGS_1_1_UDP_EXPLORER_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=1 -DOPEN_PORTS_PER_PROCESS=1  -DPROTOCOL=XTCP_PROTOCOL_UDP)
-set(APP_COMPILER_FLAGS_1_1_TCP_EXPLORER_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=1 -DOPEN_PORTS_PER_PROCESS=1  -DPROTOCOL=XTCP_PROTOCOL_TCP)
+set(APP_COMPILER_FLAGS_1_1_UDP_XMS0020_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=1 -DOPEN_PORTS_PER_PROCESS=1  -DPROTOCOL=XTCP_PROTOCOL_UDP)
+set(APP_COMPILER_FLAGS_1_1_TCP_XMS0020_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=1 -DOPEN_PORTS_PER_PROCESS=1  -DPROTOCOL=XTCP_PROTOCOL_TCP)
 # 2 processes, 2 port
-set(APP_COMPILER_FLAGS_2_2_UDP_EXPLORER_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=2 -DOPEN_PORTS_PER_PROCESS=2  -DPROTOCOL=XTCP_PROTOCOL_UDP)
-set(APP_COMPILER_FLAGS_2_2_TCP_EXPLORER_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=2 -DOPEN_PORTS_PER_PROCESS=2  -DPROTOCOL=XTCP_PROTOCOL_TCP)
+set(APP_COMPILER_FLAGS_2_2_UDP_XMS0020_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=2 -DOPEN_PORTS_PER_PROCESS=2  -DPROTOCOL=XTCP_PROTOCOL_UDP)
+set(APP_COMPILER_FLAGS_2_2_TCP_XMS0020_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=2 -DOPEN_PORTS_PER_PROCESS=2  -DPROTOCOL=XTCP_PROTOCOL_TCP)
 # 4 processes, 1 port
-# set(APP_COMPILER_FLAGS_4_1_UDP_EXPLORER_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=4 -DOPEN_PORTS_PER_PROCESS=1  -DPROTOCOL=XTCP_PROTOCOL_UDP)
-# set(APP_COMPILER_FLAGS_4_1_TCP_EXPLORER_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=4 -DOPEN_PORTS_PER_PROCESS=1  -DPROTOCOL=XTCP_PROTOCOL_TCP)
+# set(APP_COMPILER_FLAGS_4_1_UDP_XMS0020_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=4 -DOPEN_PORTS_PER_PROCESS=1  -DPROTOCOL=XTCP_PROTOCOL_UDP)
+# set(APP_COMPILER_FLAGS_4_1_TCP_XMS0020_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=4 -DOPEN_PORTS_PER_PROCESS=1  -DPROTOCOL=XTCP_PROTOCOL_TCP)
 # 4 processes, 4 ports
-# set(APP_COMPILER_FLAGS_4_4_UDP_EXPLORER_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=4 -DOPEN_PORTS_PER_PROCESS=4  -DPROTOCOL=XTCP_PROTOCOL_UDP)
-# set(APP_COMPILER_FLAGS_4_4_TCP_EXPLORER_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=4 -DOPEN_PORTS_PER_PROCESS=4  -DPROTOCOL=XTCP_PROTOCOL_TCP)
+# set(APP_COMPILER_FLAGS_4_4_UDP_XMS0020_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=4 -DOPEN_PORTS_PER_PROCESS=4  -DPROTOCOL=XTCP_PROTOCOL_UDP)
+# set(APP_COMPILER_FLAGS_4_4_TCP_XMS0020_ETH     ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=4 -DOPEN_PORTS_PER_PROCESS=4  -DPROTOCOL=XTCP_PROTOCOL_TCP)
 # 1 process, 16 ports
-# set(APP_COMPILER_FLAGS_1_16_UDP_EXPLORER_ETH    ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=1 -DOPEN_PORTS_PER_PROCESS=16 -DPROTOCOL=XTCP_PROTOCOL_UDP)
-# set(APP_COMPILER_FLAGS_1_16_TCP_EXPLORER_ETH    ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=1 -DOPEN_PORTS_PER_PROCESS=16 -DPROTOCOL=XTCP_PROTOCOL_TCP)
+# set(APP_COMPILER_FLAGS_1_16_UDP_XMS0020_ETH    ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=1 -DOPEN_PORTS_PER_PROCESS=16 -DPROTOCOL=XTCP_PROTOCOL_UDP)
+# set(APP_COMPILER_FLAGS_1_16_TCP_XMS0020_ETH    ${COMPILER_FLAGS_COMMON} -DREFLECT_PROCESSES=1 -DOPEN_PORTS_PER_PROCESS=16 -DPROTOCOL=XTCP_PROTOCOL_TCP)
 
 set(APP_XSCOPE_SRCS         config.xscope)
