--- tests/xtcp_event_driven/config.xscope
@@ -1,11 +1,0 @@
-<xSCOPEconfig ioMode="basic" enabled="true">
-    <!-- For example: -->
-    <!-- <Probe name="Probe Name" type="CONTINUOUS" datatype="UINT" units="Value" enabled="true"/> -->
-    <!-- From the target code, call: xscope_int(PROBE_NAME, value); -->
-
-    <Probe name="TOCLIENT" 		type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>
-    <Probe name="CLIENT" 	 	type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>
-    <Probe name="FROM_CLIENT" 	type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>
-    <Probe name="TO_XCOREDEV" 	type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>
-    <Probe name="TX_BUFFER" 	type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>
-</xSCOPEconfig>
