+++ tests/xtcp_xk_eth_316_dual_basic/config.xscope
@@ -1,0 +1,2 @@
+<xSCOPEconfig ioMode="basic" enabled="true">
+</xSCOPEconfig>
