+++ tests/xtcp_xk_evk_xe216_basic/config.xscope
@@ -1,0 +1,2 @@
+<xSCOPEconfig ioMode="basic" enabled="true">
+</xSCOPEconfig>
